+++ CutTheRope/ctr_commons/HLiftScrollbar.cs
@@ -26,10 +26,10 @@ namespace CutTheRope.ctr_commons
                 lift.parentAnchor = 17;
                 lift.anchor = 18;
                 lift.minX = 1f;
-                lift.maxX = (float)width - lift.minX;
+                lift.maxX = width - lift.minX;
                 lift.liftDelegate = new Lift.PercentXY(percentXY);
                 int num = 45;
-                lift.setTouchIncreaseLeftRightTopBottom((float)num, (float)num, -5f, 10f);
+                lift.setTouchIncreaseLeftRightTopBottom(num, num, -5f, 10f);
                 addChild(lift);
                 spointsNum = 0;
                 spoints = null;
