+++ CutTheRope/iframework/visual/Button.cs
@@ -65,7 +65,7 @@ namespace CutTheRope.iframework.visual
             {
                 return pointInRect(tx, ty, drawX + forcedTouchZone.x - num, drawY + forcedTouchZone.y - num, forcedTouchZone.w + num * 2f, forcedTouchZone.h + num * 2f);
             }
-            return pointInRect(tx, ty, drawX - touchLeftInc - num, drawY - touchTopInc - num, (float)width + (touchLeftInc + touchRightInc) + num * 2f, (float)height + (touchTopInc + touchBottomInc) + num * 2f);
+            return pointInRect(tx, ty, drawX - touchLeftInc - num, drawY - touchTopInc - num, width + (touchLeftInc + touchRightInc) + num * 2f, height + (touchTopInc + touchBottomInc) + num * 2f);
         }
 
         public virtual void setState(BUTTON_STATE s)
